ensure replicaCount doesn't get overridden (#23)
* ensure replicaCount doesn't get overridden

This is important when autoscaler or lingo is used. Without this change,
helm upgrade would scale down to 1 or whatever the replicaCount is set
to. This allows users to remove replicaCount from their values file and
let an autoscaler take care of it.

The default value of K8s Deployment.replicas is 1. That's why omitting
replicas if 1 is fine and why default in values.yaml is set to 1.

* upgrade chart version